Add the Category displaying (Discussions inside it)
For now, nested category crashes the app. The request only returns an ID
for them, so we need to get them from the CategoriesFragment.
Or we can display them only in the CategoriesFragment.
We need to decide: we do as the forum (no nested categories in the full
list) or we display the full list.